fix(platform): refine analyzer parser review updates

+++ openpulse-platform/src/main/java/io/github/lihanc940/openpulse/integration/analyzer/model/AnalyzerReport.java
@@ -60,10 +60,10 @@ public record AnalyzerReport(
 
     @JsonIgnoreProperties(ignoreUnknown = true)
     public record Quality(
-            long complexityScore,
-            long maintainabilityScore,
-            long documentationScore,
-            long testScore
+            int complexityScore,
+            int maintainabilityScore,
+            int documentationScore,
+            int testScore
     ) {
     }
